Refactor: moved dto packages
