+++ src/main/java/com/zerobase/hseungho/restaurantreservation/service/dto/user/SignUp.java
@@ -1,4 +1,4 @@
-package com.zerobase.hseungho.restaurantreservation.service.dto;
+package com.zerobase.hseungho.restaurantreservation.service.dto.user;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.*;
